requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
boilerpy3==1.0.7
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
datasets==2.19.1
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
gmpy2==2.1.5
google-ai-generativelanguage==0.6.4
google-ai-haystack==1.0.0
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
grpcio-tools==1.64.0
h11==0.14.0
h2==4.1.0
haystack-ai==2.1.2
haystack-bm25==1.0.2
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface_hub==0.23.2
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
lazy-imports==0.3.1
MarkupSafe==2.1.5
monotonic==1.5
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
openai==1.30.3
packaging==24.0
pandas==2.2.2
Pillow==9.4.0
pip==24.0
portalocker==2.8.2
posthog==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.2
pydantic_core==2.18.3
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.9.1
qdrant-haystack==3.7.0
regex==2024.5.15
requests==2.32.2
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sympy==1.12
tenacity==8.3.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.1
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
haystack_ai